add fp16 support

+++ mmdet/core/loss/losses.py
@@ -6,15 +6,14 @@ import torch.nn.functional as F
 def weighted_nll_loss(pred, label, weight, avg_factor=None):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
-    raw = F.nll_loss(pred, label, reduction='none')
+    raw = F.nll_loss(pred.float(), label, reduction='none')
     return torch.sum(raw * weight)[None] / avg_factor
 
 
-def weighted_cross_entropy(pred, label, weight, avg_factor=None,
-                           reduce=True):
+def weighted_cross_entropy(pred, label, weight, avg_factor=None, reduce=True):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
-    raw = F.cross_entropy(pred, label, reduction='none')
+    raw = F.cross_entropy(pred.float(), label, reduction='none')
     if reduce:
         return torch.sum(raw * weight)[None] / avg_factor
     else:
@@ -25,7 +24,7 @@ def weighted_binary_cross_entropy(pred, label, weight, avg_factor=None):
     if avg_factor is None:
         avg_factor = max(torch.sum(weight > 0).float().item(), 1.)
     return F.binary_cross_entropy_with_logits(
-        pred, label.float(), weight.float(),
+        pred.float(), label.float(), weight.float(),
         reduction='sum')[None] / avg_factor
 
 
@@ -35,12 +34,12 @@ def sigmoid_focal_loss(pred,
                        gamma=2.0,
                        alpha=0.25,
                        reduction='mean'):
-    pred_sigmoid = pred.sigmoid()
+    pred_sigmoid = pred.float().sigmoid()
     pt = (1 - pred_sigmoid) * target + pred_sigmoid * (1 - target)
     weight = (alpha * target + (1 - alpha) * (1 - target)) * weight
     weight = weight * pt.pow(gamma)
     loss = F.binary_cross_entropy_with_logits(
-        pred, target, reduction='none') * weight
+        pred.float(), target, reduction='none') * weight
     reduction_enum = F._Reduction.get_enum(reduction)
     # none: 0, mean:1, sum: 2
     if reduction_enum == 0:
@@ -68,7 +67,7 @@ def weighted_sigmoid_focal_loss(pred,
 def mask_cross_entropy(pred, target, label):
     num_rois = pred.size()[0]
     inds = torch.arange(0, num_rois, dtype=torch.long, device=pred.device)
-    pred_slice = pred[inds, label].squeeze(1)
+    pred_slice = pred[inds, label].squeeze(1).float()
     return F.binary_cross_entropy_with_logits(
         pred_slice, target, reduction='mean')[None]
 
@@ -76,7 +75,7 @@ def mask_cross_entropy(pred, target, label):
 def smooth_l1_loss(pred, target, beta=1.0, reduction='mean'):
     assert beta > 0
     assert pred.size() == target.size() and target.numel() > 0
-    diff = torch.abs(pred - target)
+    diff = torch.abs(pred.float() - target)
     loss = torch.where(diff < beta, 0.5 * diff * diff / beta,
                        diff - 0.5 * beta)
     reduction_enum = F._Reduction.get_enum(reduction)
