+++ mmdet/models/bbox_heads/bbox_head.py
@@ -110,14 +110,15 @@ class BBoxHead(nn.Module):
                        rescale=False,
                        cfg=None):
         if isinstance(cls_score, list):
-            cls_score = sum(cls_score) / float(len(cls_score))
-        scores = F.softmax(cls_score, dim=1) if cls_score is not None else None
+            cls_score = sum(cls_score).float() / float(len(cls_score))
+        scores = F.softmax(
+            cls_score.float(), dim=1) if cls_score is not None else None
 
         if bbox_pred is not None:
-            bboxes = delta2bbox(rois[:, 1:], bbox_pred, self.target_means,
-                                self.target_stds, img_shape)
+            bboxes = delta2bbox(rois[:, 1:], bbox_pred.float(),
+                                self.target_means, self.target_stds, img_shape)
         else:
-            bboxes = rois[:, 1:]
+            bboxes = rois[:, 1:].float()
             # TODO: add clip here
 
         if rescale:
